requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
google-auth==2.35.0
